+++ yggdrasil/communication/CommBase.py
@@ -909,7 +909,7 @@ class CommBase(tools.YggClass):
     #         out += '[%s]' % model_name
     #     return out
         
-    def printStatus(self, *args, level='info', **kwargs):
+    def printStatus(self, *args, level='info', return_str=False, **kwargs):
         r"""Print status of the communicator."""
         nindent = kwargs.get('nindent', 0)
         lines, prefix = self.get_status_message(*args, **kwargs)
@@ -917,6 +917,8 @@ class CommBase(tools.YggClass):
             lines.append('%sWork comms:' % prefix)
             for v in self._work_comms.values():
                 lines += v.get_status_message(nindent=nindent + 1)[0]
+        if return_str:
+            return '\n'.join(lines)
         getattr(self, level)('\n'.join(lines))
 
     @property