+++ yggdrasil/drivers/ConnectionDriver.py
@@ -647,7 +647,7 @@ class ConnectionDriver(Driver):
 
     @run_remotely
     def printStatus(self, beg_msg='', end_msg='',
-                    verbose=False):
+                    verbose=False, return_str=False):
         r"""Print information on the status of the ConnectionDriver.
 
         Arguments:
@@ -656,6 +656,8 @@ class ConnectionDriver(Driver):
             verbose (bool, optional): If True, the status of
                 individual comms will be displayed. Defaults to
                 False.
+            return_str (bool, optional): If True, the message string is
+                returned. Defaults to False.
 
         """
         msg = beg_msg
@@ -672,10 +674,14 @@ class ConnectionDriver(Driver):
             if self.close_state:
                 msg += '%-30s' % ('close state: ' + self.close_state)
         msg += end_msg
-        print(msg)
+        if not return_str:
+            print(msg)
         if verbose:
-            self.icomm.printStatus()
-            self.ocomm.printStatus()
+            i_msg = self.icomm.printStatus(return_str=return_str)
+            o_msg = self.ocomm.printStatus(return_str=return_str)
+            if return_str:
+                msg += '\n%s\n%s' % (i_msg, o_msg)
+        return msg
 
     @run_remotely
     def confirm_input(self, timeout=None):