+++ yggdrasil/tests/test_tools.py
@@ -293,6 +293,7 @@ class TestYggClass(YggTestClass):
         except Exception:
             self.instance.exception(1)
         self.instance.printStatus()
+        self.instance.printStatus(return_str=True)
         self.instance.special_debug(1)
         self.instance.suppress_special_debug = True
         self.instance.special_debug(1)